move mixins to a separate module
